Add files via upload
Your target here is to build prediction model so that you are able to predict which consumer is more likely to dispute the resolution of a complaint i.e. you need to make predictions for the "Consumer disputed" column.
All the column names are self-explanatory. You need to build your model on train data. Test data does not have the response column "Consumer disputed", you need to predict those values and submit it in a csv format.
Here you work on creating the machine learning models and choosing the one which gives the best performance.
To get a passing grade in this project you need to have the AUC score of at least 0.52 for your test data predictions.
Note: Output of test data is provided in sample_submission.csv

Task1: Build the model for trained data.
Task2: Update the existing notebook to improve the performance